+++ src/main/java/com/dbms/Residence/Finder/models/Property.java
@@ -12,7 +12,7 @@ public class Property {
     @Id
     @GeneratedValue
     @Column(name = "locality_id")
-    private int localityId;
+    private Long localityId;
 
     private String name;
